requirements.txt
datetime 
pandas
feedparser
requests
mysql-connector-python

